requirements.txt
requests
azure.common
pprint